+++ crates/matrix-sdk/src/room/mod.rs
@@ -158,7 +158,10 @@ use crate::{
     BaseRoom, Client, Error, HttpResult, Result, RoomState, TransmissionProgress,
 };
 #[cfg(feature = "e2e-encryption")]
-use crate::{crypto::types::events::CryptoContextInfo, encryption::backups::BackupState};
+use crate::{
+    crypto::types::events::CryptoContextInfo, encryption::backups::BackupState,
+    encryption::futures::ShareRoomHistory,
+};
 
 pub mod edit;
 pub mod futures;
@@ -1798,6 +1801,20 @@ impl Room {
         }
 
         Ok(())
+    }
+
+    /// Share any shareable E2EE history in this room with the given recipient,
+    /// as per [MSC4268].
+    ///
+    /// # Panics
+    ///
+    /// Panics if the OlmMachine hasn't been set up on the client.
+    ///
+    /// [MSC4268]: https://github.com/matrix-org/matrix-spec-proposals/pull/4268
+    #[cfg(feature = "e2e-encryption")]
+    #[instrument(skip_all, fields(room_id = ?self.room_id(), ?user_id))]
+    pub fn share_history<'a>(&'a self, user_id: &UserId) -> ShareRoomHistory<'a> {
+        ShareRoomHistory::new(self, user_id.to_owned())
     }
 
     /// Wait for the room to be fully synced.